+++ bundle/default-bundle/pom.xml
@@ -76,7 +76,7 @@
         <groupId>org.codehaus.mojo</groupId>
         <artifactId>exec-maven-plugin</artifactId>
         <configuration>
-          <mainClass>io.camunda.connector.runtime.ConnectorRuntimeApplication</mainClass>
+          <mainClass>io.camunda.connector.runtime.app.TestConnectorRuntimeApplication</mainClass>
         </configuration>
       </plugin>
        <plugin>
@@ -131,7 +131,7 @@
                   <resource>META-INF/io.netty.versions.properties</resource>
               </transformer>
               <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                <mainClass>io.camunda.connector.runtime.ConnectorRuntimeApplication</mainClass>
+                <mainClass>io.camunda.connector.runtime.app.TestConnectorRuntimeApplication</mainClass>
               </transformer>
             </transformers>
           </configuration>
