+++ spring-boot-starter-connectors/src/main/resources/META-INF/spring.factories
