+++ spring-boot-starter-connectors/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 #server.port=9898
-camunda.connector.polling.enabled=false
+camunda.connector.polling.enabled=true
 camunda.connector.polling.interval=5000
 
 camunda.connector.webhook.enabled=false
@@ -30,5 +30,3 @@ operate.client.enabled=true
 #camunda.operate.client.client-secret=xxx
 
 # See io.camunda.connector.inbound.operate.OperateClientFactory for more details on config Options
-
-debug=true
