+++ spring-boot-starter-connectors/src/test/java/io/camunda/connector/runtime/ObjectMapperSerializationTest.java
@@ -22,6 +22,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import io.camunda.connector.runtime.app.TestConnectorRuntimeApplication;
 import io.camunda.zeebe.client.api.JsonMapper;
 import io.camunda.zeebe.spring.test.ZeebeSpringTest;
 import java.time.ZoneOffset;
@@ -37,7 +38,8 @@ import org.springframework.test.util.ReflectionTestUtils;
     properties = {
       "spring.main.allow-bean-definition-overriding=true",
       "camunda.connector.polling.enabled=false"
-    })
+    },
+    classes = {TestConnectorRuntimeApplication.class})
 @ZeebeSpringTest
 public class ObjectMapperSerializationTest {
