+++ spring-boot-starter-connectors/src/test/resources/META-INF/services/io.camunda.connector.api.inbound.InboundConnectorExecutable
@@ -1,1 +1,1 @@
-io.camunda.connector.runtime.inbound.TestInboundConnector
+io.camunda.connector.runtime.app.TestInboundConnector
