+++ spring-boot-starter-connectors/src/test/resources/META-INF/services/io.camunda.connector.api.outbound.OutboundConnectorFunction
@@ -1,1 +1,1 @@
-io.camunda.connector.runtime.outbound.TestOutboundConnector
+io.camunda.connector.runtime.app.TestOutboundConnector
